Update controller and service to implement disable person

+++ RestWithAspNet10Scaffold/Repositories/IPersonRepository.cs
@@ -4,6 +4,6 @@ namespace RestWithAspNet10Scaffold.Repositories;
 
 public interface IPersonRepository : IRepository<Person>
 {
-    Person? Disable(int id);
-    Person? Enable(int id);
+    Person? Disable(long id);
+    Person? Enable(long id);
 }
